+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Add `- unreleased` along the next future version to prevent CI/CD from triggerin
 - Adds `Orbit.json` file to publishing process to be created alongside the copied `Orbit.toml` and `Orbit.lock` files at the destination directory in a channel to store additional metadata about the ip
 - Adds `ORBIT_BLUEPRINT_PLAN` environment variable to contain the current set plan during the build process
 - Adds `ORBIT` environment variable to contain the full path to the orbit binary that was used for the current build process
+- Adds `license` and `license-file` fields to manifest format
 
 ### Changes
 - Streamlines command and arguments for external programs by using single field `command` (removes `args` field from configurations of targets, protocols, and channels)